pyproject.toml
[tool.poetry]
name = "movie-reservation"
version = "0.1.0"
description = ""
authors = ["nicoache1 <nicoo.her@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
alembic = "^1.14.0"
fastapi = { extras = ["standard"], version = "^0.115.6" }
psycopg2-binary = "^2.9.10"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
bcrypt = "4.0.1"
passlib = "^1.7.4"
dependency-injector = "^4.44.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
mypy-extensions = "^1.0.0"
pydantic-settings = "^2.6.1"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"
